+++ app/code/core/Mage/GoogleCheckout/controllers/RedirectController.php
@@ -1,5 +1,22 @@
 <?php
-
+/**
+ * Magento
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@magentocommerce.com so we can send you a copy immediately.
+ *
+ * @category   Mage
+ * @package    Mage_GoogleCheckout
+ * @copyright  Copyright (c) 2004-2007 Irubin Consulting Inc. DBA Varien (http://www.varien.com)
+ * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
 
 class Mage_GoogleCheckout_RedirectController extends Mage_Core_Controller_Front_Action
 {
@@ -8,6 +25,7 @@ class Mage_GoogleCheckout_RedirectController extends Mage_Core_Controller_Front_
         $session = Mage::getSingleton('checkout/session');
 
         $api = Mage::getModel('googlecheckout/api')
+            ->setAnalyticsData($this->getRequest()->getPost('analyticsdata'))
             ->checkout($session->getQuote());
 
         $response = $api->getResponse();
@@ -16,22 +34,24 @@ class Mage_GoogleCheckout_RedirectController extends Mage_Core_Controller_Front_
             $url = Mage::getUrl('checkout/cart');
         } else {
             $url = $api->getRedirectUrl();
-            $session->setGoogleCheckoutQuoteId($session->getQuoteId());
-            $session->unsQuoteId();
+            if (Mage::getStoreConfigFlag('google/checkout/hide_cart_contents')) {
+                $session->setGoogleCheckoutQuoteId($session->getQuoteId());
+                $session->unsQuoteId();
+            }
         }
         $this->getResponse()->setRedirect($url);
     }
 
     public function cartAction()
     {
-        $session = Mage::getSingleton('checkout/session');
-
-        if ($session->getQuoteId()) {
-            $session->getQuote()->delete();
+        if (Mage::getStoreConfigFlag('google/checkout/hide_cart_contents')) {
+            $session = Mage::getSingleton('checkout/session');
+            if ($session->getQuoteId()) {
+                $session->getQuote()->delete();
+            }
+            $session->setQuoteId($session->getGoogleCheckoutQuoteId());
+            $session->unsGoogleCheckoutQuoteId();
         }
-
-        $session->setQuoteId($session->getGoogleCheckoutQuoteId());
-        $session->unsGoogleCheckoutQuoteId();
 
         $this->_redirect('checkout/cart');
     }
@@ -45,7 +65,9 @@ class Mage_GoogleCheckout_RedirectController extends Mage_Core_Controller_Front_
                 ->setIsActive(false)->save();
         }
 
-        $session->unsGoogleCheckoutQuoteId();
+        if (Mage::getStoreConfigFlag('google/checkout/hide_cart_contents')) {
+            $session->unsGoogleCheckoutQuoteId();
+        }
 
         $url = Mage::getStoreConfig('google/checkout/continue_shopping_url');
         if (empty($url)) {
